implement self-adapting heavytail

+++ EvolutionaryAlgorithm.Core/Parameters/IParameters.cs
@@ -8,6 +8,6 @@ namespace EvolutionaryAlgorithm.Core.Parameters
         public int GeneCount { get; set; }
         public int Mu { get; set; }
         public int Lambda { get; set; }
-        public int MutationRate { get; set; }
+        public double MutationRate { get; set; }
     }
 }
